+++ api/entities.go
@@ -7,15 +7,14 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+	"sync"
 
 	"github.com/reliablyhq/cli/core/entities"
+	log "github.com/sirupsen/logrus"
 )
 
 func CreateEntity(client *Client, hostname string, org string, entity entities.Entity) error {
-	path, err := requestPath(org, entity.Version(), entity.Kind())
-	if err != nil {
-		return err
-	}
+	path, _ := requestPath(org, entity.Version(), entity.Kind())
 
 	var body bytes.Buffer
 	if err := json.NewEncoder(&body).Encode(entity); err != nil {
@@ -39,6 +38,58 @@ func GetObjectiveResults(client *Client, hostname string, version string, org st
 
 	return entitiesResult, nil
 
+}
+
+func GetRelationshipGraph(client *Client, hostname, org string, m entities.Manifest) (*entities.NodeGraph, error) {
+	if len(m) == 0 {
+		return nil, errors.New("no entities found in manifest")
+	}
+
+	// TODO: by using m[0].Version() we assumes all entities in a manifest
+	// will have the same API version. This should be changed if/when the API
+	// is extended beyond v1
+	path, _ := requestPath(m[0].Version(), m[0].Kind(), org)
+
+	var body bytes.Buffer
+	if err := json.NewEncoder(&body).Encode(m); err != nil {
+		return nil, err
+	}
+
+	var g entities.NodeGraph
+	return &g, client.RESTv2(hostname,
+		http.MethodPost, path, &body, &g)
+}
+
+// SyncManifest - synchronize objectives from manifest with the entity server
+func SyncManifest(client *Client, entityHost, org string, m entities.Manifest) error {
+	var hasErr bool
+	var errchan = make(chan error, len(m))
+	var w sync.WaitGroup
+
+	for _, slo := range m {
+		w.Add(1)
+		go func(slo *entities.Objective) {
+			defer w.Done()
+			log.Debugf("syncing slo: %s", slo.Name)
+			if err := CreateEntity(client, entityHost, org, slo); err != nil {
+				errchan <- fmt.Errorf("error syncing manifest object: %s - %s", slo.Name, err)
+			}
+		}(slo)
+	}
+
+	w.Wait()
+	close(errchan)
+
+	for e := range errchan {
+		log.Debug(e)
+		hasErr = true
+	}
+
+	if hasErr {
+		return errors.New("An error occured while syncing your manifest")
+	}
+
+	return nil
 }
 
 func requestPath(org, version, kind string) (string, error) {