+++ cmd/reliably/slo/cmd.go
@@ -4,6 +4,7 @@ import (
 	"github.com/MakeNowJust/heredoc/v2"
 	"github.com/reliablyhq/cli/cmd/reliably/slo/agent"
 	init_cmd "github.com/reliablyhq/cli/cmd/reliably/slo/init"
+	"github.com/reliablyhq/cli/cmd/reliably/slo/related"
 	"github.com/reliablyhq/cli/cmd/reliably/slo/report"
 	"github.com/reliablyhq/cli/cmd/reliably/slo/sync"
 	"github.com/spf13/cobra"
@@ -20,6 +21,7 @@ func NewCommand() *cobra.Command {
 	cmd.AddCommand(init_cmd.NewCommand(nil))
 	cmd.AddCommand(report.NewCommand(nil))
 	cmd.AddCommand(sync.NewCommand(nil))
+	cmd.AddCommand(related.NewCommand(nil))
 	return &cmd
 }